requirements.txt
fastapi
uvicorn
llama-cpp-python
diffusers
torch
transformers
accelerate
sqlalchemy
httpx
duckduckgo-search
